- Rename checkpointstore to checkpoints - Rename BlobCheckpointStore to BlobStore
